+++ lib/src/internal/table_area_layout.dart
@@ -163,7 +163,6 @@ class _TableLayoutRenderBox extends RenderBox
     }
   }
 
-  //TODO replace by parent constraints?
   double? _width;
 
   set width(double? value) {